# doc/api.yml
openapi: 3.0.3
info:
  title: HEI Academic Management API
  description: >
    API for managing HEI students' grades and graduation over a 3-year program
    (PROG4 / SYS3). Covers authentication, account registration and role
    management, students, teachers, courses, groups, promotions, group
    history, course assignments, exams, grades (with history), transcripts,
    and graduation lists.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /promotions:
    get:
      tags: [Promotions]
      summary: List promotions
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: List of promotions, sorted by graduation year descending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Promotions]
      summary: Create a promotion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionCreateRequest"
      responses:
        "201":
          description: Promotion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /promotions/{id}/groups/{groupId}:
    post:
      tags: [Promotions]
      summary: Link a group to a promotion
      description: >
        Declares that this group is a valid group for this promotion. This
        link is the source of truth used to validate groupId on student
        creation, on PATCH /students/{id}/promotion-and-group, and on
        PATCH /students/{id}/group.
      parameters:
        - $ref: "#/components/parameters/PromotionIdPath"
        - $ref: "#/components/parameters/GroupIdPath"
      responses:
        "201":
          description: Group linked to promotion
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: This group is already linked to this promotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    GroupIdPath:
      name: groupId
      in: path
      required: true
      description: Group identifier
      schema:
        type: string
        format: uuid
...
    PromotionIdPath:
      name: id
      in: path
      required: true
      description: Promotion identifier
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 20
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient rights for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    Promotion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        graduationYear:
          type: integer
    PromotionCreateRequest:
      type: object
      required: [name, graduationYear]
      properties:
        name:
          type: string
        graduationYear:
          type: integer
...